core/editor3: moved comments related logic inside the reducers folder
