+++ scripts/core/editor3/reducers/comments/comments.js
