+++ scripts/core/editor3/reducers/comments/index.js
@@ -9,7 +9,6 @@ import {redrawComments, removeInlineStyles, applyInlineStyles} from './styles';
 import {getComments, addComment, replaceComments} from './store';
 import {updateComments} from './comments';
 
-// TODO(gbbr): Problems with highlighting multiple blocks with headers (not sure how to repro).
 // TODO(gbbr): Weird behaviour around copy/pasting comments around. Can this be fixed?
 // TODO(gbbr): Optimize for speed (for 'insert-characters' mostly). Perhaps we can
 // redraw only changes?
