+++ scripts/core/editor3/reducers/comments/styles.jsx
